updating

+++ files/bin/create_rancheros_vm.sh
@@ -1,26 +1,28 @@
 #!/usr/bin/env bash
 RANCHEROS_VERSION=1.4.0
 RANCHEROSISO=https://github.com/rancher/os/releases/download/v$RANCHEROS_VERSION/rancheros.iso
-
-echo "Retrieving RancherOS $RANCHEROS_VERSION ISO"
+if [ !  -f /var/lib/libvirt/images/iso/rancheros-auto.iso ]; then
+echo "ISO not found, Retrieving RancherOS $RANCHEROS_VERSION ISO and modifying for local installations."
 wget -cv $RANCHEROSISO -O /var/lib/libvirt/images/rancheros.iso
 mkdir -p /var/lib/libvirt/images/ros.tmp
 mount -o loop /var/lib/libvirt/images/rancheros.iso /var/lib/libvirt/images/ros.tmp
 cp -rf /var/lib/libvirt/images/ros.tmp /var/lib/libvirt/images/ros_iso
-sed -i .bak 's/APPEND/APPEND\ console\=ttyS0\,115200n9\ serial/' /var/lib/libvirt/images/ros_iso/boot/glocal.cfg
+cat << EOF > /var/lib/libvirt/images/ros_iso/boot/global.cfg
+APPEND rancher.autologin=tty1 rancher.autologin=ttyS0 rancher.autologin=ttyS1 rancher.autologin=ttyS1 console=tty1 console=ttyS0,115200n9 serial console=ttyS1 printk.devkmsg=on panic=10 rancher.cloud_init.datasources=[url:http://i.pxe.to/cloud-config.yml/rancheros-1.4.0-amd64.pxe_installer.sh]
+EOF
 cd /var/lib/libvirt/images/ros_iso && xorriso \
     -as mkisofs \
     -l -J -R -V "RANCHEROS" \
     -no-emul-boot -boot-load-size 4 -boot-info-table \
     -b boot/isolinux/isolinux.bin -c boot/isolinux/boot.cat \
     -isohybrid-mbr /usr/lib/ISOLINUX/isohdpfx.bin \
-    -o /var/lib/libvirt/images/iso/rancheros-modified.iso /var/lib/libvirt/images/ros_iso
-sleep 30
-
-#umount /var/lib/libvirt/images/tmp
-#rm -rf /var/lib/libvirt/images/tmp
-#rm -rf /var/lib/libvirt/images/rancheros.iso
-
+    -o /var/lib/libvirt/images/iso/rancheros-auto.iso /var/lib/libvirt/images/ros_iso
+sleep 10
+umount /var/lib/libvirt/images/ros.tmp
+rm -rf /var/lib/libvirt/images/ros.tmp
+rm -rf /var/lib/libvirt/images/rancheros.iso
+rm -rf /var/lib/libvirt/images/ros_iso
+fi
 virt-install \
         --name $1 \
         --os-type linux \
@@ -35,7 +37,7 @@ virt-install \
         --console pty,target_type=serial \
         --disk path=/var/lib/libvirt/images/$1.qcow2,format=qcow2,size=$2,bus=virtio \
         --nographics \
-        --cdrom /var/lib/libvirt/images/iso/rancheros-modified.iso
+        --cdrom /var/lib/libvirt/images/iso/rancheros-auto.iso
 
 #    --extra-args="
 #    console=ttyS0,115200n9 serial
